moving VerifInterpolator
